TextExtractor and TextBoxExtractor prototype. Some pom.xml changes.

+++ src/main/java/eu/virtualparadox/comictoolset/translator/bubbledetector/ComicBubbleBox.java
@@ -1,11 +1,11 @@
 package eu.virtualparadox.comictoolset.translator.bubbledetector;
 
-public record ComicBubbleBox(float x1, float y1, float x2, float y2, float confidence, int classId) {
-    public float width() {
+public record ComicBubbleBox(int x1, int y1, int x2, int y2, float confidence, int classId) {
+    public int width() {
         return x2 - x1;
     }
 
-    public float height() {
+    public int height() {
         return y2 - y1;
     }
 }
